INGM-679 import fixtures
